+++ x/sigs/msg_test.go
@@ -5,9 +5,10 @@ import (
 
 	"github.com/iov-one/weave"
 	"github.com/iov-one/weave/errors"
+	"github.com/iov-one/weave/weavetest"
 )
 
-func TextBumpSequenceValidate(t *testing.T) {
+func TestBumpSequenceValidate(t *testing.T) {
 	cases := map[string]struct {
 		Msg     weave.Msg
 		WantErr *errors.Error
@@ -16,13 +17,22 @@ func TextBumpSequenceValidate(t *testing.T) {
 			Msg: &BumpSequenceMsg{
 				Metadata:  &weave.Metadata{Schema: 1},
 				Increment: 1,
+				User:      weavetest.NewCondition().Address(),
 			},
 			WantErr: nil,
+		},
+		"missing user": {
+			Msg: &BumpSequenceMsg{
+				Metadata:  &weave.Metadata{Schema: 1},
+				Increment: 1,
+			},
+			WantErr: errors.ErrEmpty,
 		},
 		"missing metadata": {
 			Msg: &BumpSequenceMsg{
 				Metadata:  nil,
 				Increment: 1,
+				User:      weavetest.NewCondition().Address(),
 			},
 			WantErr: errors.ErrMetadata,
 		},
@@ -30,6 +40,7 @@ func TextBumpSequenceValidate(t *testing.T) {
 			Msg: &BumpSequenceMsg{
 				Metadata:  &weave.Metadata{Schema: 1},
 				Increment: 0,
+				User:      weavetest.NewCondition().Address(),
 			},
 			WantErr: errors.ErrMsg,
 		},
@@ -37,6 +48,7 @@ func TextBumpSequenceValidate(t *testing.T) {
 			Msg: &BumpSequenceMsg{
 				Metadata:  &weave.Metadata{Schema: 1},
 				Increment: 1001,
+				User:      weavetest.NewCondition().Address(),
 			},
 			WantErr: errors.ErrMsg,
 		},